dbdk/ai_based_question_recommendation_model/fm-batch/requirements.txt
aiohttp==3.7.3
aiohttp-cors==0.7.0
aioredis==1.3.1
async-timeout==3.0.1
attrs==20.3.0
blessings==1.7
cachetools==4.2.1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
colorama==0.4.4
colorful==0.5.4
filelock==3.0.12
google-api-core==1.26.0
google-auth==1.27.0
googleapis-common-protos==1.52.0
gpustat==0.6.0
grpcio==1.35.0
hiredis==1.1.0
idna==2.10
importlib-metadata==3.4.0
joblib==1.0.1
jsonschema==3.2.0
msgpack==1.0.2
multidict==5.1.0
mysql-connector-python==8.0.22
numpy==1.19.4
nvidia-ml-py3==7.352.0
opencensus==0.7.12
opencensus-context==0.1.2
packaging==20.9
pandas==1.2.0
prometheus-client==0.9.0
protobuf==3.14.0
psutil==5.8.0
py-spy==0.3.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyodbc==4.0.30
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.4.1
ray==1.1.0
redis==3.5.3
requests==2.25.1
rsa==4.7.1
scikit-learn==0.24.1
scipy==1.5.4
sentry-sdk==0.19.5
six==1.15.0
threadpoolctl==2.1.0
typing-extensions==3.7.4.3
urllib3==1.26.2
yarl==1.6.3
zipp==3.4.0
